🐛 서버바 navigate 시 버튼 색상 및 모양 변경 (#19)
